# API_Specification.yaml
swagger: '2.0'
info:
  title: 'Delilah Restó'
  description: 'Una API que permite obtener, modificar y añadir información al sistema de pedidos de la aplicación Delilah Restó. Los clientes pueden registrarse en la plataforma, ver el listado de productos y realizar pedidos. Los administradores del sitio pueden consultar toda la información de los pedidos y actualizar su estado.'
  version: '1.0.0'
  contact:
    email: 'calcetin.izquierdo@gmail.com'
  license:
    name: 'MIT License'
    url: 'https://opensource.org/licenses/MIT'
host: 'localhost:3000'
basePath: '/api/v1'
...
schemes:
  - 'https'
  - 'http'
paths:
  /users:
    get:
      tags:
        - 'usuarios'
      summary: 'Obtiene la información de todos los usuarios registrados.'
      description: ''
      operationId: 'getUsers'
      produces:
        - 'application/json'
      parameters:
        - in: header
          name: 'auth-token'
          description: 'JWT válido de administrador que se genera al iniciar sesión.'
          type: 'string'
          format: 'jwt'
          required: true
      responses:
        200:
          description: 'Arreglo de usuarios.'
          schema:
            type: 'array'
            items:
              type: 'object'
              properties:
                id:
                  type: 'integer'
                nombre:
                  type: 'string'
                telefono:
                  type: 'string'
                  minLength: 10
                  maxLength: 10
                email:
                  type: 'string'
                  format: 'email'
                direccion:
                  type: 'string'
        401:
          description: 'Acceso denegado. No se ha proporcionado un JWT o no corresponde a un administrador.'
        500:
          description: 'Error interno.'
    post:
      tags:
        - 'usuarios'
      summary: 'Registra un nuevo usuario.'
      description: ''
      operationId: 'addUser'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'body'
          description: 'Objeto de tipo usuario con toda la información requerida.'
          required: true
          schema:
            $ref: '#/definitions/Usuario'
      responses:
        200:
          description: 'Usuario creado exitosamente.'
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
              data:
                type: 'object'
                properties:
                  nombre:
                    type: 'string'
                  direccion:
                    type: 'string'
                  telefono:
                    type: 'string'
                    minLength: 10
                    maxLength: 10
                  email:
                    type: 'string'
                    format: 'email'
        400:
          description: 'Algún campo no tiene el formato indicado.'
        409:
          description: 'El email o el teléfono ya está registrado.'
        500:
          description: 'Error interno.'
  /users/myUser:
    get:
      tags:
        - usuarios
      summary: 'Obtiene la información del usuario de la sesión actual.'
      description: ''
      operationId: 'getUser'
      produces:
        - 'application/json'
      parameters:
        - in: header
          name: 'auth-token'
          type: 'string'
          format: 'jwt'
          description: 'JWT de usuario.'
          required: true
      responses:
        200:
          description: 'Información del usuario que corresponde a la sesión actual.'
          schema:
            type: 'object'
            properties:
              nombre:
                type: 'string'
              direccion:
                type: 'string'
              telefono:
                type: 'string'
                minLength: 10
                maxLength: 10
              email:
                type: 'string'
                format: 'email'
        404:
          description: 'Usuario no encontrado.'
        409:
          description: 'Acceso denegado. El token no es válido o no se ha proporcionado.'
        500:
          description: 'Error interno.'
    patch:
      tags:
        - usuarios
      summary: 'Actualiza la información del usuario de la sesión actual.'
      description: ''
      operationId: 'updateUser'
      consumes:
        - 'application/json'
      produces:
        - 'application/json'
      parameters:
        - in: header
          name: 'auth-token'
          type: 'string'
          format: 'jwt'
          description: 'JWT de usuario.'
          required: true
        - in: body
          name: 'body'
          description: 'Objeto con los campos que se desean actualizar.'
          required: true
          schema:
            $ref: '#/definitions/Usuario'
      responses:
        200:
          description: 'Usuario actualizado.'
          schema:
            type: 'object'
            properties:
              message:
                type: 'string'
              data:
                type: 'object'
                properties:
                  nombre:
                    type: 'string'
                  direccion:
                    type: 'string'
                  telefono:
                    type: 'string'
                    minLength: 10
                    maxLength: 10
                  email:
                    type: 'string'
                    format: 'email'
        400:
          description: 'Request inválido. No se enviaron campos o no corresponden al formato.'
        401:
          description: 'Acceso denegado. El token no es válido o no se ha proporcionado.'
        409:
          description: 'Nada que actualizar. Es probable que la información corresponda a la ya almacenada.'
        500:
          description: 'Error interno.'
...
definitions:
  Usuario:
    type: 'object'
    properties:
      nombre:
        type: 'string'
      apellido_p:
        type: 'string'
      apellido_m:
        type: 'string'
      telefono:
        type: 'string'
        minLength: 10
        maxLength: 10
      direccion:
        type: 'string'
      email:
        type: 'string'
        format: 'email'
      pass:
        type: 'string'
        minLength: 6
        maxLength: 15
...